requirements.txt
requests==2.31.0
pandas==2.1.4
numpy==1.24.3
python-dotenv==1.0.0
supabase==2.0.0
sqlalchemy==2.0.23
psycopg2-binary==2.9.9
geopy==2.4.1
scikit-learn==1.3.2
matplotlib==3.8.2
seaborn==0.13.0
schedule==1.2.0
python-dateutil==2.8.2
